extracted long-polling mode into a new module
