+++ examples/telegram-bot-sample-webhook/pom.xml
@@ -20,8 +20,8 @@
             <artifactId>telegram-bot-spring-boot-starter</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-web</artifactId>
+            <groupId>io.github.ksilisk</groupId>
+            <artifactId>telegram-bot-webhook</artifactId>
         </dependency>
     </dependencies>
 </project>
