+++ pom.xml
@@ -72,8 +72,9 @@
 
     <modules>
         <module>examples</module>
-        <module>telegram-bot-webhook</module>
         <module>telegram-bot-core</module>
+        <module>telegram-bot-webhook</module>
+        <module>telegram-bot-long-polling</module>
         <module>telegram-bot-spring-boot-autoconfigure</module>
         <module>telegram-bot-spring-boot-starter</module>
     </modules>
@@ -100,6 +101,11 @@
             <dependency>
                 <groupId>io.github.ksilisk</groupId>
                 <artifactId>telegram-bot-webhook</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.github.ksilisk</groupId>
+                <artifactId>telegram-bot-long-polling</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
