+++ telegram-bot-long-polling/pom.xml
@@ -10,8 +10,8 @@
         <version>${revision}</version>
     </parent>
 
-    <artifactId>telegram-bot-spring-boot-autoconfigure</artifactId>
-    <name>Telegram Bot Spring Boot Autoconfigure</name>
+    <artifactId>telegram-bot-long-polling</artifactId>
+    <name>Telegram Bot Long-Polling</name>
 
     <dependencies>
         <dependency>
@@ -19,8 +19,9 @@
             <artifactId>telegram-bot-core</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.github.ksilisk</groupId>
-            <artifactId>telegram-bot-long-polling</artifactId>
+            <groupId>jakarta.validation</groupId>
+            <artifactId>jakarta.validation-api</artifactId>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -28,28 +29,9 @@
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-configuration-processor</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-validation</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework</groupId>
-            <artifactId>spring-web</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>jakarta.servlet</groupId>
-            <artifactId>jakarta.servlet-api</artifactId>
-            <optional>true</optional>
-        </dependency>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
+
 </project>
