+++ telegram-bot-long-polling/src/main/java/io/ksilisk/telegrambot/longpolling/poller/DefaultUpdatePoller.java
@@ -1,12 +1,13 @@
-package io.ksilisk.telegrambot.core.poller;
+package io.ksilisk.telegrambot.longpolling.poller;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.GetUpdates;
 import com.pengrad.telegrambot.response.GetUpdatesResponse;
 import io.ksilisk.telegrambot.core.delivery.UpdateDelivery;
 import io.ksilisk.telegrambot.core.executor.TelegramBotExecutor;
-import io.ksilisk.telegrambot.core.properties.LongPollingProperties;
-import io.ksilisk.telegrambot.core.store.OffsetStore;
+import io.ksilisk.telegrambot.core.poller.UpdatePoller;
+import io.ksilisk.telegrambot.longpolling.properties.LongPollingProperties;
+import io.ksilisk.telegrambot.longpolling.store.OffsetStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
