+++ telegram-bot-spring-boot-autoconfigure/src/main/java/io/ksilisk/telegrambot/autoconfigure/TelegramBotAutoConfiguration.java
@@ -6,9 +6,8 @@ import io.ksilisk.telegrambot.autoconfigure.config.dispatch.DispatcherConfigurat
 import io.ksilisk.telegrambot.autoconfigure.config.dispatch.ExceptionHandlerConfiguration;
 import io.ksilisk.telegrambot.autoconfigure.config.dispatch.NoMatchStrategyConfiguration;
 import io.ksilisk.telegrambot.autoconfigure.config.dispatch.RoutingConfiguration;
-import io.ksilisk.telegrambot.autoconfigure.config.transport.LongPollingConfiguration;
+import io.ksilisk.telegrambot.autoconfigure.config.transport.LongPollingMissingIngressConfiguration;
 import io.ksilisk.telegrambot.autoconfigure.config.transport.TelegramClientConfiguration;
-import io.ksilisk.telegrambot.autoconfigure.config.transport.WebhookConfiguration;
 import io.ksilisk.telegrambot.autoconfigure.config.transport.WebhookMissingIngressConfiguration;
 import io.ksilisk.telegrambot.autoconfigure.properties.TelegramBotProperties;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
@@ -20,15 +19,14 @@ import org.springframework.context.annotation.Import;
 @EnableConfigurationProperties(TelegramBotProperties.class)
 @ConditionalOnClass(TelegramBot.class)
 @Import({
+        LongPollingMissingIngressConfiguration.class,
+        WebhookMissingIngressConfiguration.class,
         TelegramClientConfiguration.class,
         NoMatchStrategyConfiguration.class,
         ExceptionHandlerConfiguration.class,
         RoutingConfiguration.class,
         DispatcherConfiguration.class,
-        DeliveryConfiguration.class,
-        LongPollingConfiguration.class,
-        WebhookConfiguration.class,
-        WebhookMissingIngressConfiguration.class
+        DeliveryConfiguration.class
 })
 public class TelegramBotAutoConfiguration {
 }
