+++ telegram-bot-spring-boot-autoconfigure/src/main/java/io/ksilisk/telegrambot/autoconfigure/config/transport/LongPollingMissingIngressConfiguration.java
@@ -1,0 +1,20 @@
+package io.ksilisk.telegrambot.autoconfigure.config.transport;
+
+import io.ksilisk.telegrambot.core.executor.TelegramBotExecutor;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingClass;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration(proxyBeanMethods = false)
+@ConditionalOnProperty(prefix = "telegram.bot", name = "mode", havingValue = "LONG_POLLING", matchIfMissing = true)
+@ConditionalOnMissingClass("io.ksilisk.telegrambot.longpolling.ingress.LongPollingUpdateIngress")
+public class LongPollingMissingIngressConfiguration {
+    @Bean
+    public TelegramBotExecutor telegramLongPollingModeButNoIngress() {
+        throw new IllegalStateException(
+                "telegram.bot.mode=LONG_POLLING, but Long-Polling ingress is not on the classpath. " +
+                        "Please add 'telegram-bot-long-polling' dependency"
+        );
+    }
+}
