+++ telegram-bot-webhook/src/main/java/io/ksilisk/telegrambot/webhook/TelegramBotWebhookAutoConfiguration.java
@@ -1,29 +1,27 @@
-package io.ksilisk.telegrambot.autoconfigure.config.transport;
+package io.ksilisk.telegrambot.webhook;
 
-import io.ksilisk.telegrambot.webhook.controller.DefaultWebhookController;
-import io.ksilisk.telegrambot.webhook.converter.TelegramUpdateHttpMessageConverter;
-import io.ksilisk.telegrambot.webhook.filter.WebhookSecretTokenFilter;
-import io.ksilisk.telegrambot.webhook.lifecycle.DefaultWebhookLifecycle;
 import io.ksilisk.telegrambot.core.delivery.UpdateDelivery;
 import io.ksilisk.telegrambot.core.executor.TelegramBotExecutor;
-import io.ksilisk.telegrambot.webhook.ingress.WebhookUpdateIngress;
-import io.ksilisk.telegrambot.webhook.ingress.DefaultWebhookUpdateIngress;
-import io.ksilisk.telegrambot.webhook.properties.WebhookProperties;
+import io.ksilisk.telegrambot.webhook.controller.DefaultWebhookController;
 import io.ksilisk.telegrambot.webhook.controller.WebhookController;
+import io.ksilisk.telegrambot.webhook.converter.TelegramUpdateHttpMessageConverter;
+import io.ksilisk.telegrambot.webhook.filter.WebhookSecretTokenFilter;
+import io.ksilisk.telegrambot.webhook.ingress.DefaultWebhookUpdateIngress;
+import io.ksilisk.telegrambot.webhook.ingress.WebhookUpdateIngress;
+import io.ksilisk.telegrambot.webhook.lifecycle.DefaultWebhookLifecycle;
 import io.ksilisk.telegrambot.webhook.lifecycle.WebhookLifecycle;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
+import io.ksilisk.telegrambot.webhook.properties.WebhookProperties;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
 
-@Configuration(proxyBeanMethods = false)
+@AutoConfiguration
 @ConditionalOnProperty(prefix = "telegram.bot", name = "mode", havingValue = "WEBHOOK")
-@ConditionalOnClass(WebhookUpdateIngress.class)
-public class WebhookConfiguration {
+public class TelegramBotWebhookAutoConfiguration {
     @Bean
     @ConfigurationProperties(prefix = "telegram.bot.webhook")
     public WebhookProperties webhookProperties() {
